requirements.txt
# ------------------------------------------
# Lucanidae Classifier Training Requirements
# ------------------------------------------

# Data Collection & Processing
pyinaturalist>=0.19.0
pandas>=2.0.0
requests>=2.31.0
tqdm>=4.65.0
pillow>=10.0.0

# Utilities
PyYAML>=6.0
scikit-learn>=1.2.0
python-dotenv>=1.0.0
wandb>=0.16.0

# Analysis & Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Deep Learning (Note: Manual install recommended for CUDA)
# torch>=2.2.0
# torchvision>=0.17.0
